requirements.txt
Django==4.2.2
celery==5.3.0
black==23.3.0
pytest-django==4.5.2
pytest-watch==4.2.0
ipython==8.14.0
ipdb==0.13.13
